backend/requirements.txt
fastapi>=0.116,<1
uvicorn>=0.35,<1
torch>=2.10,<3
torchvision>=0.25,<1
pillow>=11,<12
numpy>=2.2,<3
requests>=2.32,<3
prometheus-client>=0.21,<1
shapely>=2.0,<3
ultralytics>=8.3,<9
supervision>=0.25,<1
